Made new RepositoryFilter and let FilterableRepository use the RepositoryFilter
