--- src/Spray/PersistenceBundle/Repository/FilterableRepositoryInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Spray\PersistenceBundle\Repository;
-
-use Spray\PersistenceBundle\EntityFilter\EntityFilterInterface;
-
-/**
- * FilterableInterface
- *
- * @author MHWK
- */
-interface FilterableRepositoryInterface
-{
-    public function filter(EntityFilterInterface $filter);
-}
